copy more files

+++ system/build.sh
@@ -1,7 +1,9 @@
 mkdir -p $PREFIX/lib $PREFIX/include $PREFIX/include
-for filename in /lib/x86_64-linux-gnu/libm.so.6 /lib/x86_64-linux-gnu/libm-2.17.so
+
+
+for filename in /lib/x86_64-linux-gnu/libm.so.6 /lib/x86_64-linux-gnu/libm-2.17.so /lib64/libm.so.6 /lib64/libm-2.12.so
 do
-  cp -a $filename $PREFIX/lib
+  test -f $filename && cp -a $filename $PREFIX/lib
 done
 cp -a /usr/include/math.h  $PREFIX/include
